requirements.txt
request==2.18.4
pycrypto==2.6.1
mock==2.0.0
